requirements.txt
requests==2.28.2
python-dotenv==0.19.2
flask==2.2.5
flask-cors==3.0.10
gunicorn==20.1.0
redis==4.5.4
celery==5.2.7
pytest==7.2.2
black==22.12.0
flake8==5.0.4
mypy==0.991
